WIP Replace macro with inline function
First draft, should work.

I'm sure there is more room for simplification, but not sure if it is
worth the hassle. The problem is that we need to specify what to add
from the the views INSIDE the iterator.

Being inlined, all extra "+=" lines should nicely collapse at compile
time.